REFACTOR CODE
