+++ .env.example
@@ -1,0 +1,10 @@
+SUPABASE_URL=Dummy
+SUPABASE_ANON_KEY=Dummy
+
+DISCORD_TOKEN=Dummy
+
+RANK_CHANNEL_ID=Dummy
+ANNOUNCEMENT_CHANNEL_ID=Dummy
+GAMING_CHANNEL_ID=Dummy
+CHANNEL_ID_1=Dummy  
+CHANNEL_ID_2=Dummy
