+++ .gitignore
@@ -78,6 +78,9 @@ web_modules/
 .env.test.local
 .env.production.local
 .env.local
+.env.staging
+.env.production
+.env.backup
 
 # parcel-bundler cache (https://parceljs.org/)
 .cache
