+++ package.json
@@ -3,7 +3,10 @@
   "version": "1.0.0",
   "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node BotShiro.js",
+    "start:staging": "cross-env NODE_ENV=staging node BotShiro.js",
+    "start:prod": "cross-env NODE_ENV=production node BotShiro.js"
   },
   "keywords": [],
   "author": "",
@@ -13,5 +16,8 @@
     "@supabase/supabase-js": "^2.49.4",
     "discord.js": "^14.19.2",
     "dotenv": "^16.5.0"
+  },
+  "devDependencies": {
+    "cross-env": "^7.0.3"
   }
 }
